+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 ## Changelog
 
+### Version 0.5.0
+
+- Adds `MIHEC*` classes to support Elliptic Curve Cryptography.
+- Fix a memory leak in `MIHRSAPrivateKey` and `MIHPublicKey` (see [issue #60](https://github.com/hohl/MIHCrypto/issues/60) for details).
+
+There are no API breaks in release 0.5.0 compared to 0.4.1. Thanks to [@zty0](https://github.com/zty0) for pointing out the memory leak.
+
 ### Version 0.4.1
 
  - Adds `signWithMD5:error:` and `verifySignatureWithMD5:message:` to `MIHRSAPrivateKey` and `MIHRSAPublicKey`.